extract common code into VariantMongoAggregationReader
