+++ eva-accession-release/src/main/java/uk/ac/ebi/eva/accession/release/io/AccessionedVariantMongoReader.java
@@ -17,10 +17,6 @@
 package uk.ac.ebi.eva.accession.release.io;
 
 import com.mongodb.MongoClient;
-import com.mongodb.client.AggregateIterable;
-import com.mongodb.client.MongoCollection;
-import com.mongodb.client.MongoCursor;
-import com.mongodb.client.MongoDatabase;
 import com.mongodb.client.model.Aggregates;
 import com.mongodb.client.model.Filters;
 import org.bson.Document;
@@ -29,10 +25,6 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.batch.item.ExecutionContext;
 import org.springframework.batch.item.ItemStreamException;
-import org.springframework.batch.item.ItemStreamReader;
-import org.springframework.batch.item.NonTransientResourceException;
-import org.springframework.batch.item.ParseException;
-import org.springframework.batch.item.UnexpectedInputException;
 
 import uk.ac.ebi.eva.commons.core.models.VariantType;
 import uk.ac.ebi.eva.commons.core.models.VariantTypeToSOAccessionMap;
@@ -53,85 +45,19 @@ import static uk.ac.ebi.eva.accession.core.ISubmittedVariant.DEFAULT_ASSEMBLY_MA
 import static uk.ac.ebi.eva.accession.core.ISubmittedVariant.DEFAULT_SUPPORTED_BY_EVIDENCE;
 import static uk.ac.ebi.eva.accession.core.ISubmittedVariant.DEFAULT_VALIDATED;
 
-public class AccessionedVariantMongoReader implements ItemStreamReader<List<Variant>> {
+public class AccessionedVariantMongoReader extends VariantMongoAggregationReader {
 
     private static final Logger logger = LoggerFactory.getLogger(AccessionedVariantMongoReader.class);
 
     private static final String DBSNP_CLUSTERED_VARIANT_ENTITY = "dbsnpClusteredVariantEntity";
 
-    protected static final String DBSNP_SUBMITTED_VARIANT_ENTITY = "dbsnpSubmittedVariantEntity";
-
-    protected static final String ACCESSION_FIELD = "accession";
-
-    private static final String REFERENCE_ASSEMBLY_FIELD = "asm";
-
-    protected static final String STUDY_FIELD = "study";
-
-    private static final String CONTIG_FIELD = "contig";
-
-    private static final String START_FIELD = "start";
-
-    private static final String TYPE_FIELD = "type";
-
-    protected static final String REFERENCE_ALLELE_FIELD = "ref";
-
-    protected static final String ALTERNATE_ALLELE_FIELD = "alt";
-
-    protected static final String CLUSTERED_VARIANT_ACCESSION_FIELD = "rs";
-
-    protected static final String SS_INFO_FIELD = "ssInfo";
-
-    protected static final String VALIDATED_FIELD = "validated";
-
-    protected static final String ASSEMBLY_MATCH_FIELD = "asmMatch";
-
-    protected static final String ALLELES_MATCH_FIELD = "allelesMatch";
-
-    protected static final String SUPPORTED_BY_EVIDENCE_FIELD = "evidence";
-
-    public static final String VARIANT_CLASS_KEY = "VC";
-
-    public static final String STUDY_ID_KEY = "SID";
-
-    public static final String CLUSTERED_VARIANT_VALIDATED_KEY = "RS_VALIDATED";
-
-    public static final String SUBMITTED_VARIANT_VALIDATED_KEY = "SS_VALIDATED";
-
-    public static final String ASSEMBLY_MATCH_KEY = "ASMM";
-
-    public static final String ALLELES_MATCH_KEY = "ALMM";
-
-    public static final String SUPPORTED_BY_EVIDENCE_KEY = "LOE";
-
-    private static final String RS_PREFIX = "rs";
-
-    protected String assemblyAccession;
-
-    protected MongoClient mongoClient;
-
-    protected String database;
-
-    protected MongoCursor<Document> cursor;
-
-    public AccessionedVariantMongoReader(String assemblyAccession, MongoClient mongoClient,
-                                         String database) {
-        this.assemblyAccession = assemblyAccession;
-        this.mongoClient = mongoClient;
-        this.database = database;
+    public AccessionedVariantMongoReader(String assemblyAccession, MongoClient mongoClient, String database) {
+        super(assemblyAccession, mongoClient, database);
     }
 
     @Override
     public void open(ExecutionContext executionContext) throws ItemStreamException {
         aggregate(DBSNP_CLUSTERED_VARIANT_ENTITY);
-    }
-
-    protected void aggregate(String collectionName) {
-        MongoDatabase db = mongoClient.getDatabase(database);
-        MongoCollection<Document> collection = db.getCollection(collectionName);
-        AggregateIterable<Document> clusteredVariants = collection.aggregate(buildAggregation())
-                                                                  .allowDiskUse(true)
-                                                                  .useCursor(true);
-        cursor = clusteredVariants.iterator();
     }
 
     List<Bson> buildAggregation() {
@@ -144,11 +70,6 @@ public class AccessionedVariantMongoReader implements ItemStreamReader<List<Vari
         return aggregation;
     }
 
-    @Override
-    public List<Variant> read() throws Exception, UnexpectedInputException, ParseException, NonTransientResourceException {
-        return cursor.hasNext() ? getVariants(cursor.next()) : null;
-    }
-
     List<Variant> getVariants(Document clusteredVariant) {
         String contig = clusteredVariant.getString(CONTIG_FIELD);
         long start = clusteredVariant.getLong(START_FIELD);
@@ -159,6 +80,7 @@ public class AccessionedVariantMongoReader implements ItemStreamReader<List<Vari
 
         Map<String, Variant> variants = new HashMap<>();
         Collection<Document> submittedVariants = (Collection<Document>)clusteredVariant.get(SS_INFO_FIELD);
+
         for (Document submittedVariant : submittedVariants) {
             String reference = submittedVariant.getString(REFERENCE_ALLELE_FIELD);
             String alternate = submittedVariant.getString(ALTERNATE_ALLELE_FIELD);
@@ -167,6 +89,7 @@ public class AccessionedVariantMongoReader implements ItemStreamReader<List<Vari
             boolean allelesMatch = submittedVariant.getBoolean(ALLELES_MATCH_FIELD, DEFAULT_ALLELES_MATCH);
             boolean assemblyMatch = submittedVariant.getBoolean(ASSEMBLY_MATCH_FIELD, DEFAULT_ASSEMBLY_MATCH);
             boolean evidence = submittedVariant.getBoolean(SUPPORTED_BY_EVIDENCE_FIELD, DEFAULT_SUPPORTED_BY_EVIDENCE);
+
             VariantSourceEntry sourceEntry = buildVariantSourceEntry(study, sequenceOntology, validated,
                                                                      submittedVariantValidated, allelesMatch,
                                                                      assemblyMatch, evidence);
@@ -174,51 +97,5 @@ public class AccessionedVariantMongoReader implements ItemStreamReader<List<Vari
             addToVariants(variants, contig, start, rs, reference, alternate, sourceEntry);
         }
         return new ArrayList<>(variants.values());
-    }
-
-    protected void addToVariants(Map<String, Variant> variants, String contig, long start, long rs, String reference,
-                               String alternate, VariantSourceEntry sourceEntry) {
-        String variantId = (contig + "_" + start + "_" + reference + "_" + alternate).toUpperCase();
-        if (variants.containsKey(variantId)) {
-            variants.get(variantId).addSourceEntry(sourceEntry);
-        } else {
-            long end = calculateEnd(reference, alternate, start);
-            Variant variant = new Variant(contig, start, end, reference, alternate);
-            variant.setMainId(buildId(rs));
-            variant.addSourceEntry(sourceEntry);
-            variants.put(variantId, variant);
-        }
-    }
-
-    private long calculateEnd(String reference, String alternate, long start) {
-        long length = Math.max(reference.length(), alternate.length());
-        return start + length - 1;
-    }
-
-    protected VariantSourceEntry buildVariantSourceEntry(String study, String sequenceOntology, boolean validated,
-                                                       boolean submittedVariantValidated, boolean allelesMatch,
-                                                       boolean assemblyMatch, boolean evidence) {
-        VariantSourceEntry sourceEntry = new VariantSourceEntry(study, study);
-        sourceEntry.addAttribute(VARIANT_CLASS_KEY, sequenceOntology);
-        sourceEntry.addAttribute(STUDY_ID_KEY, study);
-        sourceEntry.addAttribute(CLUSTERED_VARIANT_VALIDATED_KEY, Boolean.toString(validated));
-        sourceEntry.addAttribute(SUBMITTED_VARIANT_VALIDATED_KEY, Boolean.toString(submittedVariantValidated));
-        sourceEntry.addAttribute(ALLELES_MATCH_KEY, Boolean.toString(allelesMatch));
-        sourceEntry.addAttribute(ASSEMBLY_MATCH_KEY, Boolean.toString(assemblyMatch));
-        sourceEntry.addAttribute(SUPPORTED_BY_EVIDENCE_KEY, Boolean.toString(evidence));
-        return sourceEntry;
-    }
-
-    protected String buildId(long rs) {
-        return RS_PREFIX + rs;
-    }
-
-    @Override
-    public void close() throws ItemStreamException {
-        cursor.close();
-    }
-
-    @Override
-    public void update(ExecutionContext executionContext) throws ItemStreamException {
     }
 }
